refactor(voice): extract TTSPlaybackEndpoint to tts_playback.py (issue #94 phase 3)
把「播放端点」职责从 GPTSoVITSTTSProvider 抽到新模块 tts_playback.py：
- TTSPlaybackEndpoint(QObject)：QMediaPlayer/AudioSinkPlayer 播放、播放完成回调、
  fallback timeout 兜底、临时 wav 清理，以及合成结果入队的全部信号/slot
- 端点是协调器的子 QObject（parented），随协调器一起 moveToThread 到 UI 主线程，
  播放器绝不移出主线程；合成 daemon 线程经 sink（deliver_audio/deliver_prepared →
  _audio_ready 等信号）以 queued 连接把结果投回 UI 线程，语义不变
- error_occurred 仍在协调器（PetWindow 连接），由端点 error_occurred re-emit
- 协调器瘦身为「装配 + 委托」：speak_prepared/warm_up_playback/discard_prepared/
  close 委托端点；合成队列 sink 改为端点；播放状态/信号/常量全部移出协调器
- 迁 _load_qt_multimedia/_create_audio_sink_player + 播放常量；tts.py re-export
  TTSPlaybackEndpoint 与兜底常量供既有测试导入
- 重写播放段测试指向 provider._playback，Qt 桩 monkeypatch 改 tts_playback；
  修正 service_state timeout 测试的 time 补丁指向 tts_service

回归：非崩溃运行 tests/unit + tests/ui 全绿（988 passed）。tests/ui 退出阶段约 1/3
概率的 native access violation 是早于本阶段就存在的 daemon 线程/Qt 析构竞态
（已 git stash 回 commit 4 对比验证同样复现），非本提交回归。
